+++ design/src/main/kotlin/uk/gov/govuk/design/ui/model/List.kt
@@ -50,6 +50,6 @@ sealed interface InternalLinkListItemModel {
 
 data class ProgressBarListItemUiModel(
     val topText: AccessibleString,
-    val percentage: Int,
+    val percentage: Float,
     val bottomText: AccessibleString
 )
